Remove test endpoint, and it's tests
